+++ src/parts/redis.py
@@ -1,5 +1,5 @@
-from troposphere import Ref, Template, Parameter, Output
-from troposphere.elasticache import CacheCluster, SubnetGroup, SecurityGroup
+from troposphere import Ref, Template, Parameter, Output, GetAtt
+from troposphere.elasticache import CacheCluster, SubnetGroup, SecurityGroup, SecurityGroupIngress
 from dcicutils.cloudformation_utils import camelize
 
 from .network import C4NetworkExports, C4Network
@@ -16,9 +16,10 @@ class C4RedisExports(C4Exports):
 
 class C4Redis(C4Part):
     """ Builds the Redis cluster and associated resources """
-
+    STACK_NAME_TOKEN = 'redis'
+    STACK_TITLE_TOKEN = 'Redis'
     DEFAULT_ENGINE = 'redis'
-    DEFAULT_ENGINE_VERSION = '7.x'
+    DEFAULT_ENGINE_VERSION = '7.0'
     DEFAULT_NODE_COUNT = 1
     DEFAULT_CACHE_NODE_TYPE = 'cache.t4g.small'
     NETWORK_EXPORTS = C4NetworkExports()
@@ -37,7 +38,6 @@ class C4Redis(C4Part):
 
         # Build Redis Cluster
         template.add_resource(self.build_redis_subnet_group())
-        template.add_resource(self.build_redis_security_group())
         cluster = template.add_resource(self.build_redis_cache_cluster())
         template.add_output(self.output_redis_endpoint(cluster))
         return template
@@ -47,7 +47,7 @@ class C4Redis(C4Part):
         """ Outputs the endpoint URL """
         env_name = ConfigManager.get_config_setting(Settings.ENV_NAME)
         return Output(
-            f'{env_name}RedisCacheClusterEndpoint',
+            f'{camelize(env_name)}RedisCacheClusterEndpoint',
             Value=Ref(cluster),
             Description='Endpoint to connect to Redis'
         )
@@ -63,45 +63,26 @@ class C4Redis(C4Part):
                 for subnet_key in C4NetworkExports.PRIVATE_SUBNETS[:int(ConfigManager.get_config_setting(
                     Settings.SUBNET_PAIR_COUNT, default=2))]
             ],
-            Tags=self.tags.cost_tag_array(),
-        )
-
-    def build_redis_security_group(self) -> SecurityGroup:
-        """ Builds a security group for use with Redis """
-        env_name = ConfigManager.get_config_setting(Settings.ENV_NAME)
-        return SecurityGroup(
-            f'{camelize(env_name)}SecurityGroup',
-            GroupDescription=f'Security group for Redis cache cluster associated with {env_name}',
-            VpcId=Ref(self.NETWORK_EXPORTS.import_value(C4NetworkExports.VPC)),
-            SecurityGroupIngress=[
-                {
-                    'IpProtocol': 'tcp',
-                    'FromPort': '6379',
-                    'ToPort': '6379',
-                    'CidrIp': C4Network.CIDR_BLOCK
-                }
-            ],
-            Tags=self.tags.cost_tag_array()
+            Tags=self.tags.cost_tag_obj(),
         )
 
     def build_redis_cache_cluster(self) -> CacheCluster:
         """ Builds a Redis cluster in the ElastiCache paradigm """
         env_name = ConfigManager.get_config_setting(Settings.ENV_NAME)
         logical_id = self.name.logical_id(self.redis_cluster_name(env_name))
+        # TODO: add back encryption options once they are supported.
         return CacheCluster(
             logical_id,
             AutoMinorVersionUpgrade=True,
             ClusterName=logical_id,
             Engine=self.DEFAULT_ENGINE,
             EngineVersion=ConfigManager.get_config_setting(Settings.REDIS_ENGINE_VERSION,
-                                                           default=self.DEFAULT_ENGINE),
+                                                           default=self.DEFAULT_ENGINE_VERSION),
             NumCacheNodes=ConfigManager.get_config_setting(Settings.REDIS_NODE_COUNT,
                                                            default=self.DEFAULT_NODE_COUNT),
             CacheNodeType=ConfigManager.get_config_setting(Settings.REDIS_NODE_TYPE,
                                                            default=self.DEFAULT_CACHE_NODE_TYPE),
-            Description=f'Redis cluster for {env_name}',
-            CacheSubnetGroup=Ref(self.build_redis_subnet_group()),
-            TransitEncryptionEnabled=True,
-            VpcSecurityGroupIds=self.build_redis_security_group(),
-            Tags=self.tags.cost_tag_array(),
+            CacheSubnetGroupName=Ref(self.build_redis_subnet_group()),
+            VpcSecurityGroupIds=[self.NETWORK_EXPORTS.import_value(C4NetworkExports.APPLICATION_SECURITY_GROUP)],
+            Tags=self.tags.cost_tag_obj(),
         )